requirements.txt
Flask
torch
nltk
sacrebleu
pyttsx3
matplotlib
Pillow
rouge-score
meteor